add rainbow, shader, and default functionality
Not happy with shader implementation as I used a dodgy work-around for
overwriting rainbow-colored squares when in shader mode. If I were to do
this again I would spend more time on this to get a cleaner
implementation. However, a far as I know it works and does not break
anything else.